refactor(android): folder restructuring
+ moved files / micro fixes related.
