+++ apps/android/app/src/main/java/at/aau/serg/android/ui/screens/leaderboard/LeaderboardViewModel.kt
@@ -3,7 +3,7 @@ package at.aau.serg.android.ui.screens.leaderboard
 import at.aau.serg.android.network.RetrofitProvider
 import at.aau.serg.android.network.leaderboard.LeaderboardAPI
 import at.aau.serg.android.network.leaderboard.LeaderboardService
-import at.aau.serg.android.viewmodel.BaseViewModel
+import at.aau.serg.android.core.viewmodel.BaseViewModel
 import at.aau.serg.android.util.DispatcherProvider
 import at.aau.serg.android.util.DefaultDispatcherProvider
 import kotlinx.coroutines.flow.MutableStateFlow
