+++ apps/android/app/src/main/java/at/aau/serg/android/ui/screens/settings/components/SettingsTopBar.kt
@@ -4,7 +4,7 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import at.aau.serg.android.ui.general_components.BackButton
+import at.aau.serg.android.ui.components.BackButton
 
 @Composable
 fun SettingsTopBar(onBack: () -> Unit) {
